Indicate LaTeX tags and page numbers for references
References parsing is moved to get-ref-completions.coffee. The new module uses
an .aux file to extract references, which is much faster, then scanning all tex
files. Also, it is possible to extract information about tags and pages.

Drawback? The completions do not work if the tex file is not complied. But this
should not be a big deal, as we use this package to compile tex files anyway.